+++ yasha/cli.py
@@ -28,6 +28,7 @@ import encodings
 
 import click
 from click import ClickException
+from jinja2.exceptions import UndefinedError as JinjaUndefinedError
 
 from . import yasha
 from .tests import TESTS
@@ -97,6 +98,11 @@ def load_extensions(file):
             if issubclass(attr, Extension):
                 classes.append(attr)
 
+    import jinja2.defaults
+    for name, obj in inspect.getmembers(module):
+        if name in jinja2.defaults.__all__:
+            setattr(jinja2.defaults, name, obj)
+
     try:
         TESTS.update(module.TESTS)
     except AttributeError:
@@ -129,10 +135,15 @@ def load_extensions(file):
 @click.option("--no-trim-blocks", is_flag=True, help="Load Jinja with trim_blocks=False.")
 @click.option("--no-lstrip-blocks", is_flag=True, help="Load Jinja with lstrip_blocks=False.")
 @click.option("--keep-trailing-newline", is_flag=True, help="Load Jinja with keep_trailing_newline=True.")
+@click.option("--mode", type=click.Choice(['pedantic', 'debug']), help="In pedantic mode Yasha becomes extremely picky on templates, e.g. undefined variables will raise an error. In debug mode undefined variables will print as is.")
 @click.option("-M", is_flag=True, help="Outputs Makefile compatible list of dependencies. Doesn't render the template.")
 @click.option("-MD", is_flag=True, help="Creates Makefile compatible .d file alongside the rendered template.")
 @click.option('--version', is_flag=True, callback=print_version, expose_value=False, is_eager=True, help="Print version and exit.")
-def cli(template_variables, template, output, variables, extensions, encoding, include_path, no_variable_file, no_extension_file, no_trim_blocks, no_lstrip_blocks, keep_trailing_newline, m, md):
+def cli(
+        template_variables, template, output, variables, extensions,
+        encoding, include_path, no_variable_file, no_extension_file,
+        no_trim_blocks, no_lstrip_blocks, keep_trailing_newline,
+        mode, m, md):
     """Reads the given Jinja TEMPLATE and renders its content
     into a new file. For example, a template called 'foo.c.j2'
     will be written into 'foo.c' in case the output file is not
@@ -207,10 +218,11 @@ def cli(template_variables, template, output, variables, extensions, encoding, i
         tests=TESTS,
         filters=FILTERS,
         classes=CLASSES,
+        mode=mode,
         trim_blocks=not no_trim_blocks,
         lstrip_blocks=not no_lstrip_blocks,
         keep_trailing_newline=keep_trailing_newline
-    )
+   )
 
     # Get template
     if template.name == "<stdin>":
@@ -226,6 +238,9 @@ def cli(template_variables, template, output, variables, extensions, encoding, i
     context.update(yasha.parse_cli_variables(template_variables))
 
     # Finally render template and save it
-    t_stream = t.stream(context)
-    t_stream.enable_buffering(size=5)
-    t_stream.dump(output, encoding=yasha.ENCODING)
+    try:
+        t_stream = t.stream(context)
+        t_stream.enable_buffering(size=5)
+        t_stream.dump(output, encoding=yasha.ENCODING)
+    except JinjaUndefinedError as e:
+        click.echo("UndefinedError: {}".format(e), err=True)